Fix the previous rendering problem, got a new one now
What was the problem?

The formula to detect which pixel to draw (`const bool isOn` line @ Machine.cpp) was bad. It has been replaced with a valid one.
0x80 is the most significant bit value of a byte.

What is the problem now?

The problem now is that sometimes PointsToDraw returns empty vector.
The code of this function seems good, so it may be a side-effect... That's gonna be a pain to debug