Create a target chip8emu_core
I create this target because I want to
create a target chip8dump, which will be
an app used to dump memory and at each
instruction call.
This will help me to debug the emulator.

xmake seems buggy here, I'll merge it into
wip when VS will be able again to compile
sdl2wrapper, and not to say that SDL.h is
not in the include paths, when, for some
reason, it is actually 2 times in (2.0.8 and 2.0.12, which are both versions installed by xmake on my system).